test(onboard): move pure helper coverage to modules
